Npc/Summon.GetLevel + NpcData: (sbyte) cast on NpcTemplate.GetLevel() (Java byte=signed=C# sbyte; template returns byte, sbyte-consumers need cast); clears 3 CS0266 (976->973)

+++ dotnetConversion/src/Aion.GameServer/Dataholders/NpcData.cs
@@ -53,7 +53,7 @@ public class NpcData
             {
                 NpcRating rating = npc.GetRating();
                 NpcRank rank = npc.GetRank();
-                sbyte level = npc.GetLevel();
+                sbyte level = (sbyte)npc.GetLevel();
                 StatsTemplate template = npc.GetStatsTemplate();
                 if (template.GetAttack() == 0)
                     template.SetAttack(NpcStatCalculation.CalculateStat(StatEnum.PHYSICAL_ATTACK, rating, rank, level));
